feat(navigation): add separate dropdown component
