+++ src/modules/shared/components/Navigation/NavigationDropdown/NavigationDropdown.types.ts
@@ -1,0 +1,7 @@
+import { NavigationItem } from '../Navigation.types';
+
+export interface NavigationDropdownProps {
+	items: NavigationItem[][];
+	onOpen: () => void;
+	onClose: () => void;
+}
