+++ src/modules/shared/components/Navigation/NavigationDropdown/index.ts
@@ -1,0 +1,2 @@
+export { default as NavigationDropdown } from './NavigationDropdown';
+export * from './NavigationDropdown.types';
